Refactor VgaBios to become usable by outside code.
