+++ src/Spice86.Core/Emulator/InterruptHandlers/Dos/DosInt21Handler.cs
@@ -412,22 +412,7 @@ public class DosInt21Handler : InterruptHandler {
         ushort offset = _state.DX;
         string str = GetDosString(_memory, segment, offset, '$');
 
-        // Store registers
-        ushort[] savedRegisters = {_state.ES, _state.BP, _state.AX, _state.BX, _state.CX, _state.DX};
-        
-        // Set up values for bios call
-        _state.BL = 0x0F; // white on black
-        _state.ES = segment;
-        _state.BP = offset;
-        _state.CX = (ushort)str.Length;
-        _state.AL = 0x00; // no attribute, no cursor movement
-        _state.BH = _machine.Bios.CurrentVideoPage;
-        _state.DX = _machine.Bios.GetCursorPosition(_state.BH);
-        _state.AH = 0x13; // Write string
-        _machine.CallbackHandler.Run(0x10);
-        
-        // Restore registers
-        (_state.ES, _state.BP, _state.AX, _state.BX, _state.CX, _state.DX) = (savedRegisters[0], savedRegisters[1], savedRegisters[2], savedRegisters[3], savedRegisters[4], savedRegisters[5]);
+        _machine.VgaFunctions.WriteString(str);
         
         if (_loggerService.IsEnabled(LogEventLevel.Verbose)) {
             _loggerService.Verbose("PRINT STRING: {String}", str);
