+++ src/Spice86.Core/Emulator/InterruptHandlers/Timer/TimerInt8Handler.cs
@@ -38,7 +38,7 @@ public class TimerInt8Handler : InterruptHandler {
     /// Gets or set the value of the real time clock, in ticks.
     /// </summary>
     public uint TickCounterValue {
-        get => _machine.Bios.RealTimeClock; 
-        set => _machine.Bios.RealTimeClock = value;
+        get => _machine.BiosDataArea.RealTimeClock; 
+        set => _machine.BiosDataArea.RealTimeClock = value;
     }
 }
