+++ src/Spice86.Core/Emulator/InterruptHandlers/VGA/Data/RegisterValueSet.cs
@@ -39,21 +39,21 @@ internal struct RegisterValueSet {
     private static readonly byte[] GraphicsControllerRegisterValueSet6 = {0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x05, 0x0f, 0xff};
 
     public static readonly Dictionary<int, VideoMode> VgaModes = new() {
-        [0x00] = new VideoMode(new VgaMode(MemoryModel.Text, 40, 25, 4, 9, 16, VgaBios.ColorTextSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet1, 0x67, CrtControllerRegisterValueSet1, AttributeControllerRegisterValueSet1, GraphicsControllerRegisterValueSet1),
-        [0x01] = new VideoMode(new VgaMode(MemoryModel.Text, 40, 25, 4, 9, 16, VgaBios.ColorTextSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet1, 0x67, CrtControllerRegisterValueSet1, AttributeControllerRegisterValueSet1, GraphicsControllerRegisterValueSet1),
-        [0x02] = new VideoMode(new VgaMode(MemoryModel.Text, 80, 25, 4, 9, 16, VgaBios.ColorTextSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet2, 0x67, CrtControllerRegisterValueSet2, AttributeControllerRegisterValueSet1, GraphicsControllerRegisterValueSet1),
-        [0x03] = new VideoMode(new VgaMode(MemoryModel.Text, 80, 25, 4, 9, 16, VgaBios.ColorTextSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet2, 0x67, CrtControllerRegisterValueSet2, AttributeControllerRegisterValueSet1, GraphicsControllerRegisterValueSet1),
-        [0x04] = new VideoMode(new VgaMode(MemoryModel.Cga, 320, 200, 2, 8, 8, VgaBios.ColorTextSegment), 0xFF, Palettes.Cga, SequencerRegisterValueSet3, 0x63, CrtControllerRegisterValueSet3, AttributeControllerRegisterValueSet2, GraphicsControllerRegisterValueSet2),
-        [0x05] = new VideoMode(new VgaMode(MemoryModel.Cga, 320, 200, 2, 8, 8, VgaBios.ColorTextSegment), 0xFF, Palettes.Cga, SequencerRegisterValueSet3, 0x63, CrtControllerRegisterValueSet3, AttributeControllerRegisterValueSet2, GraphicsControllerRegisterValueSet2),
-        [0x06] = new VideoMode(new VgaMode(MemoryModel.Cga, 640, 200, 1, 8, 8, VgaBios.ColorTextSegment), 0xFF, Palettes.Cga, SequencerRegisterValueSet4, 0x63, CrtControllerRegisterValueSet4, AttributeControllerRegisterValueSet3, GraphicsControllerRegisterValueSet3),
-        [0x07] = new VideoMode(new VgaMode(MemoryModel.Text, 80, 25, 4, 9, 16, VgaBios.MonochromeTextSegment), 0xFF, Palettes.Monochrome, SequencerRegisterValueSet2, 0x66, CrtControllerRegisterValueSet5, AttributeControllerRegisterValueSet4, GraphicsControllerRegisterValueSet4),
-        [0x0D] = new VideoMode(new VgaMode(MemoryModel.Planar, 320, 200, 4, 8, 8, VgaBios.GraphicsSegment), 0xFF, Palettes.Cga, SequencerRegisterValueSet5, 0x63, CrtControllerRegisterValueSet6, AttributeControllerRegisterValueSet5, GraphicsControllerRegisterValueSet5),
-        [0x0E] = new VideoMode(new VgaMode(MemoryModel.Planar, 640, 200, 4, 8, 8, VgaBios.GraphicsSegment), 0xFF, Palettes.Cga, SequencerRegisterValueSet6, 0x63, CrtControllerRegisterValueSet7, AttributeControllerRegisterValueSet5, GraphicsControllerRegisterValueSet5),
-        [0x0F] = new VideoMode(new VgaMode(MemoryModel.Planar, 640, 350, 1, 8, 14, VgaBios.GraphicsSegment), 0xFF, Palettes.Monochrome, SequencerRegisterValueSet6, 0xa3, CrtControllerRegisterValueSet8, AttributeControllerRegisterValueSet6, GraphicsControllerRegisterValueSet5),
-        [0x10] = new VideoMode(new VgaMode(MemoryModel.Planar, 640, 350, 4, 8, 14, VgaBios.GraphicsSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet6, 0xa3, CrtControllerRegisterValueSet8, AttributeControllerRegisterValueSet7, GraphicsControllerRegisterValueSet5),
-        [0x11] = new VideoMode(new VgaMode(MemoryModel.Planar, 640, 480, 1, 8, 16, VgaBios.GraphicsSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet6, 0xe3, CrtControllerRegisterValueSet9, AttributeControllerRegisterValueSet8, GraphicsControllerRegisterValueSet5),
-        [0x12] = new VideoMode(new VgaMode(MemoryModel.Planar, 640, 480, 4, 8, 16, VgaBios.GraphicsSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet6, 0xe3, CrtControllerRegisterValueSet9, AttributeControllerRegisterValueSet7, GraphicsControllerRegisterValueSet5),
-        [0x13] = new VideoMode(new VgaMode(MemoryModel.Packed, 320, 200, 8, 8, 8, VgaBios.GraphicsSegment), 0xFF, Palettes.Vga, SequencerRegisterValueSet7, 0x63, CrtControllerRegisterValueSet10, AttributeControllerRegisterValueSet9, GraphicsControllerRegisterValueSet6),
-        [0x6A] = new VideoMode(new VgaMode(MemoryModel.Planar, 800, 600, 4, 8, 16, VgaBios.GraphicsSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet6, 0xe3, CrtControllerRegisterValueSet11, AttributeControllerRegisterValueSet7, GraphicsControllerRegisterValueSet5)
+        [0x00] = new VideoMode(new VgaMode(MemoryModel.Text, 40, 25, 4, 9, 16, VgaConstants.ColorTextSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet1, 0x67, CrtControllerRegisterValueSet1, AttributeControllerRegisterValueSet1, GraphicsControllerRegisterValueSet1),
+        [0x01] = new VideoMode(new VgaMode(MemoryModel.Text, 40, 25, 4, 9, 16, VgaConstants.ColorTextSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet1, 0x67, CrtControllerRegisterValueSet1, AttributeControllerRegisterValueSet1, GraphicsControllerRegisterValueSet1),
+        [0x02] = new VideoMode(new VgaMode(MemoryModel.Text, 80, 25, 4, 9, 16, VgaConstants.ColorTextSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet2, 0x67, CrtControllerRegisterValueSet2, AttributeControllerRegisterValueSet1, GraphicsControllerRegisterValueSet1),
+        [0x03] = new VideoMode(new VgaMode(MemoryModel.Text, 80, 25, 4, 9, 16, VgaConstants.ColorTextSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet2, 0x67, CrtControllerRegisterValueSet2, AttributeControllerRegisterValueSet1, GraphicsControllerRegisterValueSet1),
+        [0x04] = new VideoMode(new VgaMode(MemoryModel.Cga, 320, 200, 2, 8, 8, VgaConstants.ColorTextSegment), 0xFF, Palettes.Cga, SequencerRegisterValueSet3, 0x63, CrtControllerRegisterValueSet3, AttributeControllerRegisterValueSet2, GraphicsControllerRegisterValueSet2),
+        [0x05] = new VideoMode(new VgaMode(MemoryModel.Cga, 320, 200, 2, 8, 8, VgaConstants.ColorTextSegment), 0xFF, Palettes.Cga, SequencerRegisterValueSet3, 0x63, CrtControllerRegisterValueSet3, AttributeControllerRegisterValueSet2, GraphicsControllerRegisterValueSet2),
+        [0x06] = new VideoMode(new VgaMode(MemoryModel.Cga, 640, 200, 1, 8, 8, VgaConstants.ColorTextSegment), 0xFF, Palettes.Cga, SequencerRegisterValueSet4, 0x63, CrtControllerRegisterValueSet4, AttributeControllerRegisterValueSet3, GraphicsControllerRegisterValueSet3),
+        [0x07] = new VideoMode(new VgaMode(MemoryModel.Text, 80, 25, 4, 9, 16, VgaConstants.MonochromeTextSegment), 0xFF, Palettes.Monochrome, SequencerRegisterValueSet2, 0x66, CrtControllerRegisterValueSet5, AttributeControllerRegisterValueSet4, GraphicsControllerRegisterValueSet4),
+        [0x0D] = new VideoMode(new VgaMode(MemoryModel.Planar, 320, 200, 4, 8, 8, VgaConstants.GraphicsSegment), 0xFF, Palettes.Cga, SequencerRegisterValueSet5, 0x63, CrtControllerRegisterValueSet6, AttributeControllerRegisterValueSet5, GraphicsControllerRegisterValueSet5),
+        [0x0E] = new VideoMode(new VgaMode(MemoryModel.Planar, 640, 200, 4, 8, 8, VgaConstants.GraphicsSegment), 0xFF, Palettes.Cga, SequencerRegisterValueSet6, 0x63, CrtControllerRegisterValueSet7, AttributeControllerRegisterValueSet5, GraphicsControllerRegisterValueSet5),
+        [0x0F] = new VideoMode(new VgaMode(MemoryModel.Planar, 640, 350, 1, 8, 14, VgaConstants.GraphicsSegment), 0xFF, Palettes.Monochrome, SequencerRegisterValueSet6, 0xa3, CrtControllerRegisterValueSet8, AttributeControllerRegisterValueSet6, GraphicsControllerRegisterValueSet5),
+        [0x10] = new VideoMode(new VgaMode(MemoryModel.Planar, 640, 350, 4, 8, 14, VgaConstants.GraphicsSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet6, 0xa3, CrtControllerRegisterValueSet8, AttributeControllerRegisterValueSet7, GraphicsControllerRegisterValueSet5),
+        [0x11] = new VideoMode(new VgaMode(MemoryModel.Planar, 640, 480, 1, 8, 16, VgaConstants.GraphicsSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet6, 0xe3, CrtControllerRegisterValueSet9, AttributeControllerRegisterValueSet8, GraphicsControllerRegisterValueSet5),
+        [0x12] = new VideoMode(new VgaMode(MemoryModel.Planar, 640, 480, 4, 8, 16, VgaConstants.GraphicsSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet6, 0xe3, CrtControllerRegisterValueSet9, AttributeControllerRegisterValueSet7, GraphicsControllerRegisterValueSet5),
+        [0x13] = new VideoMode(new VgaMode(MemoryModel.Packed, 320, 200, 8, 8, 8, VgaConstants.GraphicsSegment), 0xFF, Palettes.Vga, SequencerRegisterValueSet7, 0x63, CrtControllerRegisterValueSet10, AttributeControllerRegisterValueSet9, GraphicsControllerRegisterValueSet6),
+        [0x6A] = new VideoMode(new VgaMode(MemoryModel.Planar, 800, 600, 4, 8, 16, VgaConstants.GraphicsSegment), 0xFF, Palettes.Ega, SequencerRegisterValueSet6, 0xe3, CrtControllerRegisterValueSet11, AttributeControllerRegisterValueSet7, GraphicsControllerRegisterValueSet5)
     };
 }
